LayoutTests/imported/w3c/web-platform-tests/content-security-policy/securitypolicyviolation/source-file-data-scheme-expected.txt


FAIL Violations from data:-URL scripts have a sourceFile of 'data' assert_equals: expected 16 but got 21
